fix portability concerns with size_t printfs
Apparently size_t is a concern for printf, as %lu may not be the correct size
to printf as on all platforms. There are two options to address this:
 - cast size_t to unsigned long
 - use %zu

This commit uses the former option, as the second option is apparently also
nonstandard for compilers, and also is dependent on C99. Casting seems to be
the "most" portable option, other than not using size_t, printf, or probably C.

+++ src/esl.c
@@ -91,7 +91,7 @@ next_esl_from_buffer (const uint8_t *buf, size_t buf_size, const uint8_t **esl,
   if (left < sizeof (sv_esl_t))
     {
       prlog (PR_ERR, "Not enough space left for an ESL when unpacking buffer: %lu bytes remain\n",
-             left);
+             (unsigned long) left);
       return SV_BUF_INSUFFICIENT_DATA;
     }
 
@@ -102,13 +102,13 @@ next_esl_from_buffer (const uint8_t *buf, size_t buf_size, const uint8_t **esl,
   if (claimed_size < sizeof (sv_esl_t))
     {
       prlog (PR_ERR, "ESL's claimed size is too small for the fixed contents: %lu bytes\n",
-             claimed_size);
+             (unsigned long) claimed_size);
       return SV_ESL_SIZE_INVALID;
     }
   else if (claimed_size > left)
     {
       prlog (PR_ERR, "ESL's claimed size is bigger than the data left in the buffer: %lu vs %lu\n",
-             claimed_size, left);
+             (unsigned long) claimed_size, (unsigned long) left);
       return SV_BUF_INSUFFICIENT_DATA;
     }
 
@@ -153,13 +153,13 @@ next_esl_from_buffer (const uint8_t *buf, size_t buf_size, const uint8_t **esl,
   if (claimed_size < sizeof (sv_esl_t))
     {
       prlog (PR_ERR, "ESL's claimed size is too small for the fixed contents: %lu bytes\n",
-             claimed_size);
+             (unsigned long) claimed_size);
       return SV_ESL_SIZE_INVALID;
     }
   else if (claimed_size > left)
     {
       prlog (PR_ERR, "ESL's claimed size is bigger than the data left in the buffer: %lu vs %lu\n",
-             claimed_size, left);
+             (unsigned long) claimed_size, (unsigned long) left);
       return SV_BUF_INSUFFICIENT_DATA;
     }
